Read only users - Secure CrudApi (#873)
* POC : Read only roles - Secure CrudApi

* implement POC : read only access to api

* Fix tests

* Fix tests

* More tests

* cleanup

* Use delegate pattern

* Fix tests

* cleanup

* cleanup

* cleanup

---------

+++ examples/rally-api/grails-app/init/yakworks/rally/api/BootStrap.groovy
@@ -20,7 +20,7 @@ class BootStrap {
         AppUser.withTransaction {
             AppUser.repo.flush()
             AppUser admin = new AppUser([
-                id: (Long)6, username: "developers@9ci.com", email: "developers@9ci.com", orgId: 2
+                id: 6L, username: "developers@9ci.com", email: "developers@9ci.com", orgId: 2
             ]).persist()
             admin.addRole('ADMIN', true)
             admin.addRole('MANAGER', true)
